feat(Alexa): Add ability to create a mood entry
A MongoDB instance has been set up to store participants information and is accessible by the server by making use of Entity Framework (EF) Core.

The Participants collection holds a list of participants who themselves can hold a list of mood entries.

The Alexa Skill has been set up to include an intent for creating an entry. The intent has a dialogue interface which requires three slots to be filled: mood, rating and activity, before a request is sent to the server.

For now, a hardocded ID has been used for selecting where to insert entry data.